separate old and new HTML generation
