+++ src/talos/original_templates/variant_table.html.jinja
