+++ src/talos/original_templates/variant_table_child_row.html.jinja
