--- src/talos/split_templates/sample_index.html.jinja
@@ -1,115 +1,0 @@
-<!DOCTYPE html>
-
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <title>{{ report_title|default("Talos Report") }}</title>
-    <meta name="description" content="Talos Report">
-
-    <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/jquery.tablesorter/2.31.3/css/widget.grouping.min.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/jquery.tablesorter/2.31.3/css/theme.bootstrap_4.min.css">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.3/font/bootstrap-icons.css">
-    <style type="text/css">
-      a {text-decoration: none;}
-      .group-name {font-weight: normal;}
-    </style>
-  </head>
-
-  <body>
-    <nav class="navbar navbar-dark bg-dark justify-content-start">
-      <a class="navbar-brand p-1" href="#"> {{ report_title }}</a>
-      <span class="navbar-text">
-        {{ run_datetime }}
-      </span>
-    </nav>
-
-    {# Results #}
-    <div>
-        <a href="{{ index_path }}"> LINK TO WHOLE COHORT INDEX </a>
-
-        </br>
-        ClinvArbitration, CPG's re-aggregation of ClinVar submissions: <a href="https://github.com/populationgenomics/ClinvArbitration" target="_blank">https://github.com/populationgenomics/ClinvArbitration</a>
-        </br>
-        <i class="bi-person-check-fill text-success"> Genes in panels matched to this participant by phenotype</i>
-        </br>
-        <i class="bi-person-check-fill text-info"> Genes in panels applied to all members of this cohort by config</i>
-        </br>
-        <i class="bi-person-check-fill" style="color: #FF5733"> Genes selected by Talos, which also have a semsimian phenotype match</i>
-        </br>
-        <i class="bi bi-star-fill text-warning"> New Genes in the Mendeliome</i>
-        </br>
-        <i class="bi bi-star-fill text-primary"> New Genes in phenotype-matched panels</i>
-
-        # the table header/cols
-        <table class="table tablesorter" id="variant-table">
-            <thead>
-                <tr>
-                <th class="group-separator">Individual</th>
-                <th class="group-separator">Family</th>
-                <th class="group-separator">Source</th>
-                <th class="group-false">Variant</th>
-                <th class="group-false">Gene (MOI)</th>
-                <th data-value="Y" class="group-false">Pheno Match</th>
-                <th class="group-false">Categories</th>
-                <th class="group-false sorter-shortDate" data-date-format="yyyymmdd">First Tagged</th>
-                <th class="group-false">MANE CSQ</th>
-                <th class="group-false">ClinvArbitration</th>
-                <th class="group-false">PM5 ClinvArbitration</th>
-                <th class="group-false">Ext. Label</th>
-                <th class="group-false">Flags</th>
-                <th class="group-false">Support</th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for sample in samples if sample.variants %}
-                    {% for variant in sample.variants %}
-                        {% include "variant_table_row.html.jinja" %}
-                        {% include "variant_table_child_row.html.jinja" %}
-                    {% endfor %}
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-
-    <script src="https://cdn.jsdelivr.net/npm/jquery@3.5.1/dist/jquery.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.tablesorter/2.31.3/js/jquery.tablesorter.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.tablesorter/2.31.3/js/jquery.tablesorter.widgets.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.tablesorter/2.31.3/js/widgets/widget-grouping.min.js"></script>
-
-    <script type="text/javascript">
-      $(document).ready(function() {
-        // Tablesorter: hide child rows
-        $('.tablesorter-childRow td').hide();
-
-        // Tablesorter: configure variant table
-        $("#variant-table").tablesorter({
-          theme : "bootstrap",
-          sortList : [[0,0]],
-          widgets : [ "filter", "columns", "group" ],
-          widgetOptions : {
-            // class names added to columns when sorted
-            columns: [ "", "", "" ],
-            group_collapsible : true,
-            group_collapsed   : false,
-            group_count       : false,
-            group_separator : " "
-            }
-        })
-
-        // Tablesorter: Show/hide child rows on toggle
-        $('.tablesorter').delegate('.toggle', 'click' ,function() {
-          $(this).closest('tr').nextUntil('tr:not(.tablesorter-childRow)').find('td').toggle();
-          return false;
-        });
-
-        // Enable bootstrap tool tips
-        const tooltipTriggerList = document.querySelectorAll('[data-bs-toggle="tooltip"]')
-        const tooltipList = [...tooltipTriggerList].map(tooltipTriggerEl => new bootstrap.Tooltip(tooltipTriggerEl))
-      })
-    </script>
-  </body>
-</html>
