+++ src/talos/templates/sample_index.html.jinja
@@ -5,7 +5,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
-    <title>{{ report_title }}</title>
+    <title>{{ report_title|default("Talos Report") }}</title>
     <meta name="description" content="Talos Report">
 
     <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css">
@@ -20,7 +20,7 @@
 
   <body>
     <nav class="navbar navbar-dark bg-dark justify-content-start">
-      <a class="navbar-brand p-1" href="#"> {{ report_title|default("Talos Report") }}</a>
+      <a class="navbar-brand p-1" href="#"> {{ report_title }}</a>
       <span class="navbar-text">
         {{ run_datetime }}
       </span>
@@ -28,20 +28,22 @@
 
     {# Results #}
     <div>
+        <a href="{{ index_path }}" class="btn btn-secondary" role="button">Link back to whole-cohort page</a>
 
         </br>
         ClinvArbitration, CPG's re-aggregation of ClinVar submissions: <a href="https://github.com/populationgenomics/ClinvArbitration" target="_blank">https://github.com/populationgenomics/ClinvArbitration</a>
         </br>
-        <i class="bi-person-check-fill text-success"> Genes in panels matched to this participant by phenotype</i>
+        <i class="bi-person-check-fill text-success">Phenotype match: Panel-based</i>
         </br>
-        <i class="bi-person-check-fill text-info"> Genes in panels applied to all members of this cohort by config</i>
+        <i class="bi-person-check-fill text-info">Phenotype match: Cohort-based</i>
         </br>
-        <i class="bi-person-check-fill" style="color: #FF5733"> Genes selected by Talos, which also have a semsimian phenotype match</i>
+        <i class="bi-person-check-fill" style="color: #FF5733">Phenotype match: HPO-based</i>
         </br>
         <i class="bi bi-star-fill text-warning"> New Genes in the Mendeliome</i>
         </br>
         <i class="bi bi-star-fill text-primary"> New Genes in phenotype-matched panels</i>
 
+        # the table header/cols
         <table class="table tablesorter" id="variant-table">
             <thead>
                 <tr>
@@ -65,6 +67,7 @@
                 {% for sample in samples if sample.variants %}
                     {% for variant in sample.variants %}
                         {% include "variant_table_row.html.jinja" %}
+                        {% include "variant_table_child_row.html.jinja" %}
                     {% endfor %}
                 {% endfor %}
             </tbody>
