+++ src/talos/templates/variant_table_row.html.jinja
